aws-stack/server/requirements.txt
langchain==0.2.6
langgraph==0.1.4
tavily-python==0.3.3
openai==1.35.7
anthropic==0.30.0
python-dotenv==1.0.1
llama-index==0.10.43
llama-index-core==0.10.43
llama-index-readers-web==0.1.17
llama-index-llms-openai==0.1.22
reportlab==4.2.0
chainlit==1.1.301
Jinja2==3.1.4
google-auth==2.29.0
google-api-python-client==2.131.0
googlemaps==4.10.0
fastapi==0.110.3
uvicorn==0.25.0
fitz==0.0.1.dev2
pymupdf4llm==0.0.5
playwright==1.44.0
geopy==2.4.1
geocoder==1.38.1
PyMuPDF==1.24.7